+++ crates/tests-e2e/src/bitcoind.rs
@@ -1,0 +1,65 @@
+use anyhow::bail;
+use anyhow::Result;
+use reqwest::Client;
+use reqwest::Response;
+use serde::Deserialize;
+use std::time::Duration;
+
+/// A wrapper over the bitcoind HTTP API
+///
+/// It does not aim to be complete, functionality will be added as needed
+pub struct Bitcoind {
+    client: Client,
+    host: String,
+}
+
+impl Bitcoind {
+    pub fn new(client: Client) -> Self {
+        let host = "http://localhost:8080/bitcoin".to_string();
+        Self { client, host }
+    }
+
+    /// Instructs `bitcoind` to generate to address.
+    pub async fn mine(&self, n: u16) -> Result<()> {
+        #[derive(Deserialize, Debug)]
+        struct BitcoindResponse {
+            result: String,
+        }
+
+        let response: BitcoindResponse = self
+            .client
+            .post(&self.host)
+            .body(r#"{"jsonrpc": "1.0", "method": "getnewaddress", "params": []}"#.to_string())
+            .send()
+            .await?
+            .json()
+            .await?;
+
+        self.client
+            .post(&self.host)
+            .body(format!(
+                r#"{{"jsonrpc": "1.0", "method": "generatetoaddress", "params": [{}, "{}"]}}"#,
+                n, response.result
+            ))
+            .send()
+            .await?;
+
+        // For the mined blocks to be picked up by the subsequent wallet syncs
+        tokio::time::sleep(Duration::from_secs(5)).await;
+
+        Ok(())
+    }
+
+    pub async fn post(&self, endpoint: &str, body: Option<String>) -> Result<Response> {
+        let mut builder = self.client.post(endpoint.to_string());
+        if let Some(body) = body {
+            builder = builder.body(body);
+        }
+        let response = builder.send().await?;
+
+        if !response.status().is_success() {
+            bail!(response.text().await?)
+        }
+        Ok(response)
+    }
+}