+++ mobile/native/src/ln_dlc/mod.rs
@@ -17,12 +17,16 @@ use anyhow::Result;
 use bdk::bitcoin::secp256k1::rand::thread_rng;
 use bdk::bitcoin::secp256k1::rand::RngCore;
 use bdk::bitcoin::secp256k1::SecretKey;
+use bdk::bitcoin::Txid;
 use bdk::bitcoin::XOnlyPublicKey;
 use bdk::BlockTime;
 use coordinator_commons::TradeParams;
 use itertools::chain;
 use itertools::Itertools;
+use lightning::util::events::Event;
 use lightning_invoice::Invoice;
+use ln_dlc_node::node::rust_dlc_manager::subchannel::LNChannelManager;
+use ln_dlc_node::node::rust_dlc_manager::ChannelId;
 use ln_dlc_node::node::LnDlcNodeSettings;
 use ln_dlc_node::node::NodeInfo;
 use ln_dlc_node::seed::Bip39Seed;
@@ -39,8 +43,10 @@ use std::sync::Arc;
 use std::time::Duration;
 use time::OffsetDateTime;
 use tokio::runtime::Runtime;
+use tokio::sync::watch;
 use tokio::task::spawn_blocking;
 
+mod lightning_subscriber;
 mod node;
 
 static NODE: Storage<Arc<Node>> = Storage::new();
@@ -75,10 +81,29 @@ pub async fn update_node_settings(settings: LnDlcNodeSettings) {
     node.inner.update_settings(settings).await;
 }
 
-// TODO: should we also wrap the oracle as `NodeInfo`. It would fit the required attributes pubkey
-// and address.
 pub fn get_oracle_pubkey() -> XOnlyPublicKey {
     NODE.get().inner.oracle_pk()
+}
+
+pub fn get_funding_transaction(channel_id: &ChannelId) -> Result<Txid> {
+    let node = NODE.get();
+    let channel_details = node.inner.channel_manager.get_channel_details(channel_id);
+
+    let funding_transaction = match channel_details {
+        Some(channel_details) => match channel_details.funding_txo {
+            Some(funding_txo) => funding_txo.txid,
+            None => bail!(
+                "Could not find funding transaction for channel {}",
+                hex::encode(channel_id)
+            ),
+        },
+        None => bail!(
+            "Could not find channel details for {}",
+            hex::encode(channel_id)
+        ),
+    };
+
+    Ok(funding_transaction)
 }
 
 /// Lazily creates a multi threaded runtime with the the number of worker threads corresponding to
@@ -133,6 +158,8 @@ pub fn run(data_dir: String, seed_dir: String, runtime: &Runtime) -> Result<()> 
         let seed_path = seed_dir.join("seed");
         let seed = Bip39Seed::initialize(&seed_path)?;
 
+        let (event_sender, event_receiver) = watch::channel::<Option<Event>>(None);
+
         let node = Arc::new(ln_dlc_node::node::Node::new_app(
             "10101",
             network,
@@ -144,8 +171,14 @@ pub fn run(data_dir: String, seed_dir: String, runtime: &Runtime) -> Result<()> 
             seed,
             ephemeral_randomness,
             config::get_oracle_info(),
+            event_sender,
         )?);
         let node = Arc::new(Node { inner: node });
+
+        runtime.spawn({
+            let node = node.clone();
+            async move { node.listen_for_lightning_events(event_receiver).await }
+        });
 
         runtime.spawn({
             let node = node.clone();
@@ -385,7 +418,7 @@ pub fn create_invoice(amount_sats: Option<u64>) -> Result<Invoice> {
                 "http://{}/api/register_invoice/{}",
                 config::get_http_endpoint(),
                 node.inner.info.pubkey
-            )) // TODO: make host configurable
+            ))
             .send()
             .await?;